requirements.txt
flask~=3.0.0
flask-admin
flask-sqlalchemy
Flask-Security-Too
flask-login
flask-script
flask-mail
flask-wtf
flask-assets
Flask-JWT-Extended
flask-migrate
flask_cors
flask-debugtoolbar
flask-marshmallow
flask-rebar

oauth2~=1.9.0.post1
jinja2
markdown~=3.5.1
wtforms~=3.1.1
sqlalchemy~=2.0.25
marshmallow-sqlalchemy
apispec[yaml]~=6.3.1
apispec-webframeworks
marshmallow~=3.20.1

simplejson
lxml
pygments~=2.17.2
httplib2

tqdm~=4.66.1
pydot~=2.0.0
six~=1.16.0
bcrypt
python-slugify~=8.0.1
pytest
natsort~=8.4.0
psycopg2-binary
email-validator
Werkzeug~=3.0.1
alembic~=1.13.1
gunicorn
redis
huey~=2.5.0
compare50
pdfkit~=1.0.0
thefuzz~=0.20.0
MarkupSafe~=2.1.3
pytz
edit_distance
click
webassets